requirements.txt
Flask[async]
python-telegram-bot==13.7
httpx
mysql-connector
